+++ SIEFB/src/main/java/SIEFB/service/impl/PersonaServiceImpl.java
@@ -3,8 +3,6 @@ package SIEFB.service.impl;
 import java.util.List;
 
 
-
-
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,28 +30,47 @@ public class PersonaServiceImpl implements PersonaService {
 
 	    @Override
 	    public Persona guardar(Persona persona) {
+	    	// Validar si ya existe el documento
+	        if (personaRepository.existsByDocumento(persona.getDocumento())) {
+	            throw new RuntimeException("Ya existe una persona con el documento: " 
+	                                       + persona.getDocumento());
+	        }
+
 	        return personaRepository.save(persona);
 	    }
 
 	    @Override
 	    public Persona actualizar(Integer id, Persona persona) {
-	        return personaRepository.findById(id)
-	            .map(p -> {
-	                p.setDocumento(persona.getDocumento());
-	                p.setNombre(persona.getNombre());
-	                p.setDireccion(persona.getDireccion());
-	                p.setTelefono(persona.getTelefono());
-	                p.setActaMedica(persona.getActaMedica());
-	                p.setEps(persona.getEps());
-	                p.setFoto(persona.getFoto());
-	                p.setFechaNacimiento(persona.getFechaNacimiento());
-	                return personaRepository.save(p);
-	            })
-	            .orElseThrow(() -> new RuntimeException("Persona no encontrada con id: " + id));
+	    	return personaRepository.findById(id)
+	    	        .map(p -> {
+
+	    	            // Validar documento si fue cambiado
+	    	            if (!p.getDocumento().equals(persona.getDocumento()) &&
+	    	                personaRepository.existsByDocumento(persona.getDocumento())) {
+
+	    	                throw new RuntimeException("Ya existe una persona con el documento: "
+	    	                                            + persona.getDocumento());
+	    	            }
+
+	    	            p.setDocumento(persona.getDocumento());
+	    	            p.setNombre(persona.getNombre());
+	    	            p.setDireccion(persona.getDireccion());
+	    	            p.setTelefono(persona.getTelefono());
+	    	            p.setActaMedica(persona.getActaMedica());
+	    	            p.setEps(persona.getEps());
+	    	            p.setFoto(persona.getFoto());
+	    	            p.setFechaNacimiento(persona.getFechaNacimiento());
+
+	    	            return personaRepository.save(p);
+	    	        })
+	    	        .orElseThrow(() -> new RuntimeException("Persona no encontrada con id: " + id));
 	    }
 
 	    @Override
 	    public void eliminar(Integer id) {
-	        personaRepository.deleteById(id);
+	    	if (!personaRepository.existsById(id)) {
+	    	    throw new RuntimeException("Persona no encontrada con id: " + id);
+	    	}
+	    	personaRepository.deleteById(id);
 	    }
 }
